+++ src/shared/types/index.ts
@@ -1,3 +1,4 @@
 export * from './entities';
 export * from './connections';
 export * from './embeddings';
+export * from './clusters';